registry type for records
