+++ opsml/cli/api_cli.py
@@ -165,9 +165,7 @@ def list_cards(
 
     """
 
-    registry_name = getattr(RegistryTableNames, registry.upper())
-
-    if registry_name is None:
+    if registry not in ["model", "data", "pipeline", "run"]:
         raise ValueError(
             f"No registry found. Accepted values are 'model', 'data', 'pipeline', and 'run'. Found {registry}",
             registry,
@@ -186,11 +184,11 @@ def list_cards(
         "limit": limit,
         "max_date": max_date,
         "tags": tags,
-        "table_name": registry_name,
+        "registry_type": registry,
     }
     cards = api_client.list_cards(payload=payload)
 
-    table = Table(title=f"{registry_name} cards")
+    table = Table(title=f"{registry} cards")
     table.add_column("Name", no_wrap=True)
     table.add_column("Team")
     table.add_column("Date")
