+++ opsml/registry/__init__.py
@@ -1,5 +1,5 @@
 from opsml.registry.cards import AuditCard, DataCard, ModelCard, PipelineCard, RunCard
-from opsml.registry.cards.types import CardInfo, ModelCardMetadata, DataCardMetadata
+from opsml.registry.cards.types import CardInfo, ModelCardMetadata, DataCardMetadata, RegistryType
 from opsml.registry.data.splitter import DataSplit
-from opsml.registry.sql.registry import CardRegistries, CardRegistry, RegistryType
+from opsml.registry.sql.registry import CardRegistries, CardRegistry
 from opsml.registry.sql.semver import VersionType
