+++ opsml/registry/cards/types.py
@@ -12,6 +12,15 @@ from opsml.model.types import ApiDataSchemas, DataDict, ExtraOnnxArgs, OnnxModel
 logger = ArtifactLogger.get_logger()
 
 
+class RegistryType(str, Enum):
+    DATA = "data"
+    MODEL = "model"
+    RUN = "run"
+    PIPELINE = "pipeline"
+    AUDIT = "audit"
+    PROJECT = "project"
+
+
 class Metric(BaseModel):
     name: str
     value: Union[float, int]
